Make encrypting secrets with symmetric-encryption optional
Previously it was assumed you were using symmetric-encryption to encrypt
the secrets in the database for extra security, however the gem itself
wasn't actually required, leading to trouble setting things up for the
first time.

Now encryption is treated as an optional added layer of security with
instructions on how to accomplish it. This should also open the door to
using alternate encryption strategies.

This should fix #21.